+++ app/Http/Controllers/Admin/RoomController.php
@@ -24,15 +24,18 @@ class RoomController extends Controller
      */
     public function search(Request $request)
     {
+        $this->authorize('viewAny',Room::class);
         $data = $request->all();
         return $this->room_service->searchRoomService($data);
     }
     public function index()
     {
+        $this->authorize('viewAny',Room::class);
         return view('admin.rooms.index');
     }
     public function getAllRoom()
     {
+        $this->authorize('viewAny',Room::class);
         return $this->room_service->getAllRoomsService();
     }
     /**
@@ -42,8 +45,8 @@ class RoomController extends Controller
      */
     public function create()
     {
+        $this->authorize('create',Room::class);
         $room = new Room();
-
         return view('admin.rooms.store', compact('room'));
     }
 
@@ -55,7 +58,7 @@ class RoomController extends Controller
      */
     public function store(StoreRoomRequest $request)
     {
-
+        $this->authorize('create',Room::class);
         $data = $request->all();
         $data['created_by'] = Auth::user()->id;
         try {
@@ -76,8 +79,8 @@ class RoomController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('view',Room::findOrFail($id));
         $room = $this->room_service->showRoomService($id);
-     
         return view('admin.rooms.show', compact('room'));
     }
 
@@ -89,6 +92,7 @@ class RoomController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('view',Room::findOrFail($id));
         $room = $this->room_service->findRoomService($id);
         return view('admin.rooms.edit', compact('room'));
     }
@@ -102,7 +106,7 @@ class RoomController extends Controller
      */
     public function update(StoreRoomRequest $request, $id)
     {
-
+        $this->authorize('update',Room::findOrFail($id));
         try {
             $data = $request->all();
             $this->room_service->updateRoomService($data, $id);
@@ -122,6 +126,7 @@ class RoomController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete',Room::findOrFail($id));
         try {
             $this->room_service->deleteRoomService($id);
             toastr('Xóa phòng thành công', 'success', 'Thành công');
