+++ resources/views/admin/asset/_indexscript.blade.php
@@ -26,7 +26,8 @@
                         name: 'code',
                         width: '100px',
                         render: function(data, type, row) {
-                            return data ? data : '<span class="badge badge-secondary">Không có mã</span>';
+                            return data ? data :
+                                '<span class="badge badge-secondary">Không có mã</span>';
                         }
                     },
                     {
@@ -71,16 +72,27 @@
                             var edit = row.edit_url;
                             var del = row.delete_url;
                             var show = row.show_url;
-                            var buttonEdit = '<a href="' + edit +
-                                '" class="btn btn-primary"><i class="fa fa-edit "></i></a>';
-                            var buttonShow = '<a href="' + show +
-                                '" class="btn btn-info"><i class="fa fa-eye"></i></a>';
-                            var buttonDelete = '<form action="' + del +
-                                '" method="POST" style="display:inline;" class="">' +
-                                '@CSRF' +
-                                '@method('DELETE')' +
-                                '<button class="btn btn-danger " type="submit" onclick="return confirm(\'Bạn có chắc chắn ngừng hoạt động danh mục tài sản này?\')">' +
-                                '<i class="fa fa-trash"></i></button></form>';
+
+                            var buttonEdit = '';
+                            var buttonShow = '';
+                            var buttonDelete = '';
+
+                            if (edit != null) {
+                                buttonEdit = '<a href="' + edit +
+                                    '" class="btn btn-primary"><i class="fa fa-edit "></i></a>';
+                            }
+                            if (del != null) {
+                                buttonDelete = '<form action="' + del +
+                                    '" method="POST" style="display:inline;" class="">' +
+                                    '@CSRF' +
+                                    '@method('DELETE')' +
+                                    '<button class="btn btn-danger " type="submit" onclick="return confirm(\'Bạn có chắc chắn ngừng hoạt động danh mục tài sản này?\')">' +
+                                    '<i class="fa fa-trash"></i></button></form>';
+                            }
+                            if (show) {
+                                buttonShow = '<a href="' + show +
+                                    '" class="btn btn-info"><i class="fa fa-eye"></i></a>';
+                            }
                             return buttonEdit + ' ' + buttonShow + ' ' + buttonDelete
                         },
                         width: '120px'
